# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Marketing Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthSignupBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginBody"
      responses:
        "200":
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                required: [accessToken, refreshToken]

  # ── Settings ──────────────────────────────────────────────────────────────
...
  /clients/{clientId}/campaigns/{campaignId}/generate-plan:
    post:
      operationId: generateCampaignPlan
      tags: [ai-engine]
      summary: Generate an AI content plan for a campaign (creates post drafts)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePlanBody"
      responses:
        "200":
          description: Generated posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratePlanResponse"
  /clients/{clientId}/posts/{postId}/regenerate-copy:
    post:
      operationId: regeneratePostCopy
      tags: [ai-engine]
      summary: Regenerate AI caption and hashtags for an existing post
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /clients/{clientId}/posts/{postId}/generate-image:
    post:
      operationId: generatePostImage
      tags: [ai-engine]
      summary: Generate a DALL-E image for a post and save it
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated post with image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /clients/{clientId}/generate-bulk:
    post:
      operationId: bulkGeneratePosts
      tags: [ai-engine]
      summary: Bulk-generate a full week or month of content for a brand
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkGenerateBody"
      responses:
        "200":
          description: All generated posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkGenerateResponse"
components:
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        campaignId:
          type: string
        topic:
          type: string
        caption:
          type: string
        hashtags:
          type: string
        selectedImageUrl:
          type: string
        status:
          type: string
          enum: [draft, approved, scheduled, published, failed, rejected]
        scheduledAt:
          type: string
        platform:
          type: string
          enum: [instagram, facebook, twitter, linkedin, blog, newsletter]
        postType:
          type: string
          enum: [social, blog, newsletter]
        title:
          type: string
        longFormBody:
          type: string
        generationStatus:
          type: string
          enum: [generating, ready, failed]
        imagePrompt:
          type: string
        publishedAt:
          type: string
        publishedUrl:
          type: string
        publishError:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, topic, caption, status, postType, createdAt, updatedAt]
...
    AuthSignupBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password, name]
    AuthLoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
      required: [id, email, name]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [user]

    # ── Settings ──────────────────────────────────────────────────────────────
...
    GeneratePlanBody:
      type: object
      properties:
        postsCount:
          type: number
        platforms:
          type: array
          items:
            type: string
      required: []
    GeneratePlanResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        generatedCount:
          type: number
      required: [posts, generatedCount]
    BulkGenerateBody:
      type: object
      properties:
        weeks:
          type: number
        postsPerWeek:
          type: number
        platforms:
          type: array
          items:
            type: string
        campaignId:
          type: string
      required: []
    BulkGenerateResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        generatedCount:
          type: number
      required: [posts, generatedCount]